Add factor option to anchor.
Add the option to anchor an OGWidget to a point based on the screensize.
This looks very similar to what anchor already does, but allows it to work
nicely with the stretch parameter, which was very hard to do before.
It currently looks a bit weird in the inspector, but I'm not sure how to
fix that.